refactor: cast volume as float instead of int

+++ src/window.py
@@ -1037,7 +1037,7 @@ class CineWindow(Adw.ApplicationWindow):
         return True
 
     def _update_volume_icon(self):
-        volume = cast(int, self.mpv.volume)
+        volume = cast(float, self.mpv.volume)
         is_muted = self.mpv.mute
 
         if is_muted or volume == 0:
